move encryptionkm to pkg
