+++ server/storage/storage.go
@@ -19,10 +19,10 @@ import (
 	"sync/atomic"
 
 	"github.com/pingcap/kvproto/pkg/metapb"
+	"github.com/tikv/pd/pkg/encryption"
 	"github.com/tikv/pd/pkg/storage/kv"
 	"github.com/tikv/pd/pkg/utils/syncutil"
 	"github.com/tikv/pd/server/core"
-	"github.com/tikv/pd/server/encryptionkm"
 	"github.com/tikv/pd/server/storage/endpoint"
 	"go.etcd.io/etcd/clientv3"
 )
@@ -58,7 +58,7 @@ func NewStorageWithEtcdBackend(client *clientv3.Client, rootPath string) Storage
 func NewStorageWithLevelDBBackend(
 	ctx context.Context,
 	filePath string,
-	ekm *encryptionkm.KeyManager,
+	ekm *encryption.Manager,
 ) (Storage, error) {
 	return newLevelDBBackend(ctx, filePath, ekm)
 }
